Player turns added (sort of need debugging, see description)
Player turns are done. Just need to hook it up to server (waiting on
Sean for that)

BUG!!! Placing white pieces is acting weird. It won't look -right- for
pieces, and sometimes you can place a piece somewhere you can't.

Going to look into that this weekend, have other stuff to do before.