Deleted bug & more
Alright so ShowPossibleMove is bugged for diagonals. They either don't
show up at all, or show up in weird places. The previous version (that
supposedly worked) crashed when it reached edges so meh.

Also framework for winning condition is done, just need to figure out
WHERE to place it so that it works.

Where it is right now, it'll only work if there's no possible move, and
an empty space left and doesn't come up automatically. Will keep working
on it.